requirements.txt
streamlit
face_recognition
numpy
pymongo
pillow
cmake
dlib
